# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
"description": "A full-stack development basic platform using gin+vue for extremely fast development",
"title": "Gin-Vue-Admin Swagger API interface document",
        "contact": {},
        "version": "v2.6.1"
    },
    "paths": {
...
        "/authority/copyAuthority": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authority"
                ],
"summary": "copy character",
                "parameters": [
                    {
"description": "Old role id, new permission id, new permission name, new parent role id",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.SysAuthorityCopyResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
"description": "Copy role, return including system role details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.SysAuthorityResponse"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/authority/createAuthority": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authority"
                ],
"summary": "Create role",
                "parameters": [
                    {
"description": "Permission id, permission name, parent role id",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.SysAuthority"
                        }
                    }
                ],
                "responses": {
                    "200": {
"description": "Create role, return including system role details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.SysAuthorityResponse"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/authority/deleteAuthority": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authority"
                ],
"summary": "Delete role",
                "parameters": [
                    {
"description": "Delete role",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.SysAuthority"
                        }
                    }
                ],
                "responses": {
                    "200": {
"description": "Delete role",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/authority/getAuthorityList": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authority"
                ],
"summary": "Get the role list by pagination",
                "parameters": [
                    {
"description": "Page number, size of each page",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.PageInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
"description": "Get the role list in paging, and return the list, total number, page number, and number per page",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.PageResult"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/authority/setDataAuthority": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authority"
                ],
"summary": "Set role resource permissions",
                "parameters": [
                    {
"description": "Set role resource permissions",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.SysAuthority"
                        }
                    }
                ],
                "responses": {
                    "200": {
"description": "Set role resource permissions",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/authority/updateAuthority": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authority"
                ],
"summary": "Update character information",
                "parameters": [
                    {
"description": "Permission id, permission name, parent role id",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.SysAuthority"
                        }
                    }
                ],
                "responses": {
                    "200": {
"description": "Update role information, return including system role details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.SysAuthorityResponse"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.PageInfo": {
            "type": "object",
            "properties": {
                "keyword": {
"description": "Keyword",
                    "type": "string"
                },
                "page": {
"description": "Page number",
                    "type": "integer"
                },
                "pageSize": {
"description": "Size of each page",
                    "type": "integer"
                }
            }
        },
...
        "response.PageResult": {
            "type": "object",
            "properties": {
                "list": {},
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "response.SysAuthorityCopyResponse": {
            "type": "object",
            "properties": {
                "authority": {
                    "$ref": "#/definitions/system.SysAuthority"
                },
                "oldAuthorityId": {
"description": "Old role ID",
                    "type": "integer"
                }
            }
        },
        "response.SysAuthorityResponse": {
            "type": "object",
            "properties": {
                "authority": {
                    "$ref": "#/definitions/system.SysAuthority"
                }
            }
        },
...
        "system.Meta": {
            "type": "object",
            "properties": {
                "activeName": {
                    "type": "string"
                },
                "closeTab": {
"description": "Automatically close tab",
                    "type": "boolean"
                },
                "defaultMenu": {
"description": "Whether it is a basic route (under development)",
                    "type": "boolean"
                },
                "icon": {
"description": "menu icon",
                    "type": "string"
                },
                "keepAlive": {
"description": "Whether to cache",
                    "type": "boolean"
                },
                "title": {
"description": "menu name",
                    "type": "string"
                }
            }
        },
...
        "system.SysAuthority": {
            "type": "object",
            "properties": {
                "authorityId": {
"description": "Role ID",
                    "type": "integer"
                },
                "authorityName": {
"description": "role name",
                    "type": "string"
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.SysAuthority"
                    }
                },
                "createdAt": {
"description": "Creation time",
                    "type": "string"
                },
                "dataAuthorityId": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.SysAuthority"
                    }
                },
                "defaultRouter": {
"description": "Default menu (default dashboard)",
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.SysBaseMenu"
                    }
                },
                "parentId": {
"description": "Parent role ID",
                    "type": "integer"
                },
                "updatedAt": {
"description": "Update time",
                    "type": "string"
                }
            }
        },
...
        "system.SysBaseMenu": {
            "type": "object",
            "properties": {
                "ID": {
"description": "Primary key ID",
                    "type": "integer"
                },
                "authoritys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.SysAuthority"
                    }
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.SysBaseMenu"
                    }
                },
                "component": {
"description": "Corresponding front-end file path",
                    "type": "string"
                },
                "createdAt": {
"description": "Creation time",
                    "type": "string"
                },
                "hidden": {
"description": "Whether to hide in the list",
                    "type": "boolean"
                },
                "menuBtn": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.SysBaseMenuBtn"
                    }
                },
                "meta": {
"description": "Additional attributes",
                    "allOf": [
                        {
                            "$ref": "#/definitions/system.Meta"
                        }
                    ]
                },
                "name": {
"description": "route name",
                    "type": "string"
                },
                "parameters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.SysBaseMenuParameter"
                    }
                },
                "parentId": {
"description": "Parent menu ID",
                    "type": "string"
                },
                "path": {
"description": "Routing path",
                    "type": "string"
                },
                "sort": {
"description": "Sort mark",
                    "type": "integer"
                },
                "updatedAt": {
"description": "Update time",
                    "type": "string"
                }
            }
        },
        "system.SysBaseMenuBtn": {
            "type": "object",
            "properties": {
                "ID": {
"description": "Primary key ID",
                    "type": "integer"
                },
                "createdAt": {
"description": "Creation time",
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sysBaseMenuID": {
                    "type": "integer"
                },
                "updatedAt": {
"description": "Update time",
                    "type": "string"
                }
            }
        },
        "system.SysBaseMenuParameter": {
            "type": "object",
            "properties": {
                "ID": {
"description": "Primary key ID",
                    "type": "integer"
                },
                "createdAt": {
"description": "Creation time",
                    "type": "string"
                },
                "key": {
"description": "key that carries parameters in the address bar",
                    "type": "string"
                },
                "sysBaseMenuID": {
                    "type": "integer"
                },
                "type": {
"description": "Whether the parameters carried in the address bar are params or query",
                    "type": "string"
                },
                "updatedAt": {
"description": "Update time",
                    "type": "string"
                },
                "value": {
"description": "The value of the parameter carried in the address bar",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}