# server/model/system/response/sys_authority.go
package response

import "github.com/flipped-aurora/gin-vue-admin/server/model/system"

type SysAuthorityResponse struct {
	Authority system.SysAuthority `json:"authority"`
}

type SysAuthorityCopyResponse struct {
	Authority      system.SysAuthority `json:"authority"`
OldAuthorityId uint                `json:"oldAuthorityId"` // Old role ID
}